Removed hyphens from profile widget ID's
Hey Eva i'm not sure if this will solve the issue but could we try removing the hyphens from the widget ID's where the profile widget is used. I will keep an eye on your environment to make sure this change is applied and if it still isn't resolved I will raise this back to the development team.