Added role attribute
Hey Eva i've closed the other pull request as I don't believe this would have fixed your issue. I've added a role attribute to the footer so we can see if when the templates update on your environment it is actually working. If you have any questions feel free to drop me an email.